Rename file housing MlflowLogger class to remove arbitrary version-ing from prototyping. Fix import statements in logging callback module to correctly reference MlflowLogger.
